individual hpp file for main menu scene class
