+++ include/scene.hpp
@@ -43,28 +43,6 @@ public:
     static void draw(sf::RenderWindow &window);
 };
 
-class MainMenuScene : public Scene {
-private:
-    sf::Texture bigRectanglePicture;
-    sf::Sprite bigRectangle;
-
-    sf::Texture buttonSingleplayerPicture;
-    sf::Texture buttonMultiplayerPicture;
-    sf::Sprite buttonSingleplayer;
-    sf::Sprite buttonMultiplayer;
-
-    sf::Font font;
-    sf::Text text;
-    const sf::Color backgroundColor = sf::Color(125, 186, 179);
-
-public:
-    MainMenuScene();
-
-    void update(sf::Time &dTime) override;
-    void draw(sf::RenderWindow &window) override;
-    void handleInput(sf::Event &event) override;
-};
-
 }  // namespace super_hse
 
 #endif  // SCENE_HPP_
