+++ src/game.cpp
@@ -1,6 +1,7 @@
 #include "game.hpp"
 #include <memory>
 #include "scene.hpp"
+#include "main_menu_scene.hpp"
 
 
 #include <iostream>
