+++ src/level_map_scene.cpp
@@ -1,5 +1,6 @@
 #include "level_map_scene.hpp"
 #include "level_scene.hpp"
+#include "main_menu_scene.hpp"
 #include <memory>
 #include "scene.hpp"
 #include <filesystem>
